+++ src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './env';
 export * from './telegram';
 export * from './common';
+export * from './grammy-extensions';
 export type { CloudflareEnv } from './env';